Extract BaseView
